split and move components
